pro-main/server/fakecurrency/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cloudinary==1.43.0
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
mysqlclient==2.2.7
pillow==11.1.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
